% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backbone}
\alias{backbone}
\title{GBIF Taxonomic Backbone}
\format{
\subsection{\code{backbone}}{

A data frame with 30,610 rows and 3 columns:
\describe{
\item{original}{Species name in the original dataset}
\item{gbif}{Species name in GBIF backbone}
\item{status}{Status of the nomenclature}
}
}
}
\usage{
backbone
}
\description{
The GBIF taxonomic backbone for the species in the datasets obtained
running the function \code{gbif_taxonomy}.
}
\keyword{datasets}
