% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds}
\alias{birds}
\title{Body Mass for Birds.}
\format{
\subsection{\code{birds}}{

A data frame with 9,123 rows and 2 columns:
\describe{
\item{species}{Species name}
\item{mass}{Body mass (g)}
}
}
}
\source{
\url{https://doi.org/10.5061/dryad.jm63xsj7b}
}
\usage{
birds
}
\description{
A subset of data from EltonTraits, including only species with known body mass.
For details, see \url{https://doi.org/10.1890/13-1917.1}.
}
\keyword{datasets}
