% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-inference.R
\name{compare_models}
\alias{compare_models}
\title{Compare Inference Models}
\usage{
compare_models(models)
}
\arguments{
\item{models}{List of models to compare. Better if it's a named list.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Compare Inference Models
}
\examples{
\dontrun{
library(hattian)
fw <- matrix(sample(c(0, 1), 100, replace = TRUE), nrow = 10, ncol = 10)
m_prey <- matrix(runif(10, 0, 1), nrow = 10, ncol = 10)
m_pred <- matrix(runif(10, 0, 1), nrow = 10, ncol = 10, byrow = TRUE)
fit1 <- yeakel_stan(fw, m_prey, m_pred, warmup = 100, iter = 200)
fit2 <- li_stan(fw, m_prey, m_pred, warmup = 100, iter = 200)
models <- list(yeakel = fit1, li = fit2)
compare_models(models)
}

}
