% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-inference.R
\name{confusion}
\alias{confusion}
\title{Confusion Matrix}
\usage{
confusion(fw, preds, th)
}
\arguments{
\item{fw}{Adjacency matrix of the food-web.}

\item{preds}{Numeric matrix with probabilistic prediction from an inference model.}

\item{th}{Numeric value of the threshold to apply.}
}
\value{
The confusion matrix.
}
\description{
Confusion Matrix
}
