% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.tetraeu_matrix}
\alias{.tetraeu_matrix}
\title{Convert TetraEU into Matrix}
\usage{
.tetraeu_matrix()
}
\value{
A matrix version of TetraEU.
}
\description{
Convert TetraEU into Matrix
}
\details{
This function is internal and, in theory, should not be called directly.
}
