% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-inference.R
\name{find_threshold}
\alias{find_threshold}
\title{Find Optimal Threshold}
\usage{
find_threshold(fw, preds, steps = 100)
}
\arguments{
\item{fw}{Adjacency matrix of the food-web.}

\item{preds}{Numeric matrix with probabilistic prediction from an inference model.}

\item{steps}{Numeric value of the steps of threshold test.}
}
\value{
A list with the best threshold, sensitivity, specificity, and TSS statistics.
}
\description{
Find Optimal Threshold
}
\details{
The optimal threshold is the one that maximizes the sum of specificity and sensitivity.
}
