% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize.R
\name{harmonize}
\alias{harmonize}
\title{Harmonize Taxonomy}
\usage{
harmonize(backbone, datasets, species_cols = rep("species", length(datasets)))
}
\arguments{
\item{backbone}{Character vector of taxonomic backbone.}

\item{datasets}{List of datasetes to harmonize.}

\item{species_cols}{Character vector of the column names with species names.}
}
\value{
A list of the harmonized datasets.
}
\description{
Harmonize Taxonomy
}
