% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hattian-package.R
\docType{package}
\name{hattian-package}
\alias{hattian-package}
\alias{hattian}
\title{The 'hattian' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Emilio Berti \email{emilio.berti.academia@gmail.com} (\href{https://orcid.org/0000-0001-9286-011X}{ORCID})

}
