% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_li.R
\name{li_predict}
\alias{li_predict}
\title{Predict Using Li Model}
\usage{
li_predict(fit, m_prey, m_pred, samples = 300)
}
\arguments{
\item{fit}{Numeric matrix of the adjacency matrix of the food web.
Predators are in columns and prey in rows.}

\item{m_prey}{Numeric vector of the prey body mass.}

\item{m_pred}{Numeric vector of the predator body mass.}

\item{samples}{Numeric value of the number of samples to draw.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Predict Using Li Model
}
\examples{
\dontrun{
library(hattian)
m_prey <- log10(seq(1.1, 1e2, length.out = 10))
m_pred <- log10(seq(1.1, 1e2, length.out = 10))
mus <- seq(0.5, 1.5, length.out = 10)
sigmas <- seq_len(10) / 1e2
thetas <- sqrt(seq_len(10) / 10)
p <- matrix(NA, nrow = 10, ncol = 10)
for (i in seq_along(m_pred)) {
  p[, i] <- thetas[i] * exp( -(m_pred[i] / m_prey - mus[i])^2 / (2 * sigmas[i]^2))
}
fw <- matrix(rbinom(10^2, 1, p), nrow = 10, ncol = 10, byrow = TRUE)
image(fw)
fit <- li_stan(fw, m_prey, m_pred, warmup = 100, iter = 200, refresh = 10, cores = 4)
preds <- li_predict(fit, m_prey, m_pred)
}
}
