% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_li.R
\name{li_stan}
\alias{li_stan}
\title{Li et al. (2024) model.}
\usage{
li_stan(fw, m_prey, m_pred, ...)
}
\arguments{
\item{fw}{Numeric matrix of the adjacency matrix of the food web.
Predators are in columns and prey in rows.}

\item{m_prey}{Numeric matrix of the prey body mass.}

\item{m_pred}{Numeric matrix of the predator body mass.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Li et al. (2024) model.
}
