% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mammals}
\alias{mammals}
\title{Body Mass for Mammals.}
\format{
\subsection{\code{mammals}}{

A data frame with 4,041 rows and 2 columns:
\describe{
\item{species}{Species name}
\item{mass}{Body mass (g)}
}
}
}
\source{
\url{https://doi.org/10.5061/dryad.jm63xsj7b}
}
\usage{
mammals
}
\description{
A subset of data from EltonTraits, including only species with known body mass.
For details, see \url{https://doi.org/10.1890/13-1917.1}.
}
\keyword{datasets}
