% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random-forest.R
\name{random_forest}
\alias{random_forest}
\title{Random Forest model}
\usage{
random_forest(fw, masses)
}
\arguments{
\item{fw}{Numeric matrix of the adjacency matrix of the food web.
Predators are in columns and prey in rows.}

\item{masses}{Numeric vector of the species' body mass.}

\item{...}{Arguments passed to \code{randomForest} (e.g. ntree, mtry).}
}
\value{
An object of class \code{randomForest} returned by
randomForest::randomForest`
}
\description{
Random Forest model
}
