% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reptiles}
\alias{reptiles}
\title{Body Mass for Reptiles}
\format{
\subsection{\code{reptiles}}{

A data frame with 12,059 rows and 3 columns:
\describe{
\item{species}{Species name}
\item{mass}{Body mass (g)}
\item{imputed}{Was body mass imputed using major axis regression?}
}
}
}
\source{
\url{https://doi.org/10.1038/s41597-024-03079-5}
}
\usage{
reptiles
}
\description{
A subset of data from ReptTraits, oncluding only species with known body mass
or known body length, from whcih body masses was imputed.
For details, see \url{https://doi.org/10.1038/s41597-024-03079-5}.
}
\keyword{datasets}
