% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fw-utils.R
\name{show_web}
\alias{show_web}
\title{Show Web}
\usage{
show_web(fw)
}
\arguments{
\item{fw}{Numeric matrix of the adjacency matrix of the food web.
Predators are in columns and prey in rows.}
}
\description{
Show Web
}
