% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tetraeu}
\alias{tetraeu}
\title{European Tetrapod Meta-Web}
\format{
\subsection{\code{who}}{

A data frame with 60,453 rows and 2 columns:
\describe{
\item{prey}{Prey name}
\item{predator}{Predator name}
}
}
}
\source{
\url{https://doi.org/10.5061/dryad.jm63xsj7b}
}
\usage{
tetraeu
}
\description{
A subset of data from TetraEU, including only adults individuals.
For details, see \url{https://doi.org/10.1111/geb.13138}.
}
\keyword{datasets}
