% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_yeakel.R
\name{yeakel_stan}
\alias{yeakel_stan}
\title{Yeakel et al. (?) model.}
\usage{
yeakel_stan(fw, masses, ...)
}
\arguments{
\item{fw}{Numeric matrix of the adjacency matrix of the food web.
Predators are in columns and prey in rows.}

\item{masses}{Numeric vector of the species' body mass.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Yeakel et al. (?) model.
}
\examples{
\dontrun{
library(hattian)
fw <- matrix(sample(c(0, 1), 100, replace = TRUE), nrow = 10, ncol = 10)
masses <- runif(10, 0, 1)
fit <- yeakel_stan(fw, masses, warmup = 100, iter = 200)
}
}
