% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibians}
\alias{amphibians}
\title{Body Mass for Amphibians.}
\format{
\subsection{\code{amphibians}}{

A data frame with 5,227 rows and 2 columns:
\describe{
\item{species}{Species name}
\item{mass}{Body mass (g)}
\item{imputed}{Was body mass imputed using major axis regression?}
}
}
}
\source{
\url{https://doi.org/10.5061/dryad.jm63xsj7b}
}
\usage{
amphibians
}
\description{
A subset of data from AmphiBIO, including only species with known body mass
or known body length, from which body masses was imputed.
For details, see \url{https://doi.org/10.1038/sdata.2017.123}.
}
\keyword{datasets}
