% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backbone}
\alias{backbone}
\title{GBIF Taxonomic Backbone}
\format{
\subsection{\code{backbone}}{

A data frame with 13,158 rows and 2 columns:
\describe{
\item{species}{Species name}
\item{status}{Status of the nomenclature}
}
}
}
\usage{
backbone
}
\description{
The GBIF taxonomic backbone for the species in the datasets obtained
running the function \code{gbif_taxonomy}.
}
\keyword{datasets}
