% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_li.R
\name{li_stan}
\alias{li_stan}
\title{Li et al. (2024) model.}
\usage{
li_stan(fw, masses, ...)
}
\arguments{
\item{fw}{Numeric matrix of the adjacency matrix of the food web.
Predators are in columns and prey in rows.}

\item{masses}{Numeric vector of the species' body mass.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Li et al. (2024) model.
}
\examples{
\dontrun{
library(hattian)
m_prey <- log10(seq(1.1, 1e2, length.out = 10))
m_pred <- log10(seq(1.1, 1e2, length.out = 10))
mus <- seq(0.5, 1.5, length.out = 10)
sigmas <- seq_len(10) / 1e2
thetas <- sqrt(seq_len(10) / 10)
p <- matrix(NA, nrow = 10, ncol = 10)
for (i in seq_along(m_pred)) {
  p[, i] <- thetas[i] * exp( -(m_pred[i] / m_prey - mus[i])^2 / (2 * sigmas[i]^2))
}
fw <- matrix(rbinom(10^2, 1, p), nrow = 10, ncol = 10, byrow = TRUE)
image(fw)
fit <- li_stan(fw, m_prey, m_pred, warmup = 100, iter = 200, refresh = 10, cores = 4)
}
}
